Updated with 2018-2020 regulations PDF (downloaded 2018-08-14). Updated compile / target SDK version to 27 (Oreo). Gradle 4.4, Gradle plugin 3.1.4, support lib 27.1.1, play services 15.0.1 Fixed bug on Regulations fragment where disclaimer / index button was not showing the first time in.
