+++ WildlifeTracker/app/src/main/java/ca/bc/gov/fw/wildlifetracker/RegulationsDisclaimerDialogFragment.java
@@ -1,0 +1,31 @@
+package ca.bc.gov.fw.wildlifetracker;
+
+import android.app.Dialog;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.v4.app.DialogFragment;
+import android.support.v7.app.AlertDialog;
+
+public class RegulationsDisclaimerDialogFragment extends DialogFragment {
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setMessage(R.string.regs_disclaimer_message)
+                .setTitle(R.string.regs_disclaimer_title)
+                .setPositiveButton(R.string.regs_disclaimer_ok_button, null)
+                .setNegativeButton(R.string.regs_disclaimer_view_online_button, new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+                        Uri webpage = Uri.parse("http://www2.gov.bc.ca/gov/content/sports-culture/recreation/fishing-hunting/hunting/regulations-synopsis");
+                        Intent intent = new Intent(Intent.ACTION_VIEW, webpage);
+                        if (intent.resolveActivity(getActivity().getPackageManager()) != null) {
+                            startActivity(intent);
+                        }
+                    }
+                });
+        return builder.create();
+    }
+}
