+++ WildlifeTracker/build.gradle
@@ -2,10 +2,11 @@
 
 buildscript {
     repositories {
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.5.0'
+        classpath 'com.android.tools.build:gradle:3.1.4'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -13,11 +14,13 @@ buildscript {
 }
 
 ext {
-    supportLibVersion = '23.2.1'
+    supportLibVersion = '27.1.1'
+    playServicesVersion = '15.0.1'
 }
 
 allprojects {
     repositories {
+        google()
         jcenter()
     }
 }
